# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Board of Issues task management system.",
        "title": "Board Of Issues API",
        "contact": {
            "name": "Nikita Kleymenov",
            "url": "https://t.me/n1ktarchik",
            "email": "klejmenov663@email.com"
        },
        "version": "2.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/desks": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all desks where current user is a member or owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "desks"
                ],
                "summary": "Get my desks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token for authentication (format: Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of desks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_domain.Desk"
                            }
                        }
                    },
                    "400": {
                        "description": "Possible: invalid_user_id, bad_request",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth,IdempotencyKey": []
                    }
                ],
                "description": "Create a new board for tasks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "desks"
                ],
                "summary": "Create a desk",
                "parameters": [
                    {
                        "description": "Desk Info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_desks_transport_http.DeskRequestDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Unique idempotency key (UUID) to prevent duplicate processing",
                        "name": "X-Req-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token for authentication (format: Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created desk",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_domain.Desk"
                        }
                    },
                    "400": {
                        "description": "Possible: desk_name_too_short, invalid_data",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update desk name or password (must be owner)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "desks"
                ],
                "summary": "Update desk info",
                "parameters": [
                    {
                        "description": "New desk data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_desks_transport_http.DeskRequestDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer token for authentication (format: Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated desk information",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_domain.Desk"
                        }
                    },
                    "400": {
                        "description": "Possible: invalid_uuid, bad_request",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "not_an_owner",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "desk_not_found",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/desks/connect": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth,IdempotencyKey": []
                    }
                ],
                "description": "Connect current user to a desk using desk password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "desks"
                ],
                "summary": "Join a desk",
                "parameters": [
                    {
                        "description": "Desk ID and Password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_desks_transport_http.DeskRequestDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Unique idempotency key (UUID) to prevent duplicate processing",
                        "name": "X-Req-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token for authentication (format: Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: you have connected to desk",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Possible: invalid_data, bad_request",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "desk_not_found",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "already_a_member",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "N1ktarchik_Board_of_issues_internal_core_domain.Desk": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "636e856-e12b-56d9-f987-333222561234"
                },
                "name": {
                    "type": "string",
                    "example": "My Desk"
                },
                "owner_id": {
                    "type": "string",
                    "example": "832t758-a12g-47y9-i999-123456789098"
                },
                "password": {
                    "type": "string",
                    "example": "mysecretpassword"
                }
            }
        },
...
        "N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "error": {
                    "type": "string",
                    "example": "any string-code of error"
                },
                "message": {
                    "type": "string",
                    "example": "any error message"
                }
            }
        },
...
        "internal_features_desks_transport_http.DeskRequestDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "636e856-e12b-56d9-f987-333222561234"
                },
                "name": {
                    "type": "string",
                    "example": "My Desk"
                },
                "owner_id": {
                    "type": "string",
                    "example": "832t758-a12g-47y9-i999-123456789098"
                },
                "password": {
                    "type": "string",
                    "example": "mysecretpassword"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token in the format: Bearer \u003cJWT_TOKEN\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "IdempotencyKey": {
            "description": "An idempotent key to prevent duplicates (UUID)",
            "type": "apiKey",
            "name": "X-Req-Key",
            "in": "header"
        }
    }
}